First commit- bit late in development;
So far in the project, much of the boiler plate code has been completed.
There are 'objects' such as the Printer, Toolkit, Modifier and Converter
which contain and organise the functions used.

The functions which make elementary row operations which are
required to convert a matrix to RREF are complete.